% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchability.R
\name{catchability}
\alias{catchability}
\title{This function calculates the probability of entry to trap (aka catchability)}
\usage{
catchability(q0 = 0.5, qmin = 0, saturationThreshold = 5, Ct = 0)
}
\arguments{
\item{q0}{is the initial probability of entry into an empty trap which is set to 0.5 by default}

\item{qmin}{is the asymptotic minimum probability of entry which is set to 0}

\item{saturationThreshold}{is the number of lobsters in a trap at which the probability of
another lobster entering the trap drops to qmin, the default value is 5}

\item{Ct}{is the number of catch}

\item{r}{is the instantaneous rate of change in qo with respect to Ct}
}
\value{
the probability of entry to trap
}
\description{
This function calculates the probability of entry to trap (aka catchability)
}
