% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalMove.R
\name{directionalMove}
\alias{directionalMove}
\title{This function moves the lobster for a fixed length toward the trap (+some randomness)
This is performed only when a lobster is within the bait influence}
\usage{
directionalMove(
  xLobster,
  yLobster,
  dStep,
  distanceToTrap,
  xtrap = 5,
  ytrap = 5,
  radius_of_influence = 15,
  ZoI
)
}
\arguments{
\item{dStep}{is how much a lobster moves in each time step}

\item{ZoI}{is trap's zone of influence which gets updated in each timestpe}

\item{default}{value for radius of influence is 15}

\item{thetaT}{calculates angle to the trap}
}
\value{
a list of new location for each individual lobster
}
\description{
This function moves the lobster for a fixed length toward the trap (+some randomness)
This is performed only when a lobster is within the bait influence
}
