% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{dispersion}
\alias{dispersion}
\title{This function calculates the variance to mean ratio also known as dispersion index}
\usage{
dispersion(x)
}
\description{
This function calculates the variance to mean ratio also known as dispersion index
}
