% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceClosestTrap.R
\name{distanceClosestTrap}
\alias{distanceClosestTrap}
\title{This function returns id of the closest trap and the and calculates its distance}
\usage{
distanceClosestTrap(lob_loc, trap_loc)
}
\arguments{
\item{lob_loc}{is the location of lobster}

\item{trap_loc}{is the location of trap}
}
\description{
This function returns id of the closest trap and the and calculates its distance
}
