% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceToTrapCalculator.R
\name{distanceToTrapCalculator}
\alias{distanceToTrapCalculator}
\title{This function calculates the Euclidean distance between trap(s) and
individual lobsters}
\usage{
distanceToTrapCalculator(Lobster, trap = x(5, 5))
}
\description{
This function calculates the Euclidean distance between trap(s) and
individual lobsters
}
