% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomMove.R
\name{randomMove}
\alias{randomMove}
\title{The function randomly selects an angle (0:360) and moves the lobster. This is called when a lobster
is outside the area of influence.}
\usage{
randomMove(xLobster, yLobster, dStep)
}
\arguments{
\item{xLobster}{current x coordinate of the lobster}

\item{yLobster}{current y coordinate of the lobster}

\item{dStep}{fixed length of movemnet in each timestep}
}
\description{
The function randomly selects an angle (0:360) and moves the lobster. This is called when a lobster
is outside the area of influence.
}
