% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpoisD.R
\name{rpoisD}
\alias{rpoisD}
\title{This function generates either a Poisson or a negative binomial distribution of lobsters on the sea bed}
\usage{
rpoisD(n, lambda, D = 1)
}
\arguments{
\item{n}{is the number of observations}

\item{lambda}{is the mean}

\item{D}{is the dispersion index}

\item{sz}{this puts the overdispersion in terms of lambda(mean) and returns the appropriate size for Binomial distribution}
}
\value{
a vector of integers that is used as initial distribution of lobsters on the seabed.
}
\description{
This function generates either a Poisson or a negative binomial distribution of lobsters on the sea bed
}
