% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateLobsterMovement.R
\name{SimulateLobsterMovement}
\alias{SimulateLobsterMovement}
\title{This function moves lobsters based on the provided parameters. A distribution of lengths
#has been also provided which enables the user to assign length to individuals and include the
#effect of size in catchability}
\usage{
SimulateLobsterMovement(p = p, plot = F)
}
\arguments{
\item{p}{can be modified to check the impact of each factor (i.e p$smult = 0.9 sets the shrinkage factor)}

\item{plot}{is set to false by default}
}
\description{
This function moves lobsters based on the provided parameters. A distribution of lengths
#has been also provided which enables the user to assign length to individuals and include the
#effect of size in catchability
}
