% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapInPath.R
\name{trapInPath}
\alias{trapInPath}
\title{This function determines if lobster passes through a trap and determines if it is caught}
\usage{
trapInPath(loc1, loc2, trap_loc, how_close = 0.1)
}
\arguments{
\item{loc1}{is the location of lobster at the start of each time step}

\item{loc2}{is the location of lobster at the end of each time step}

\item{trap_loc}{is the location of trap}

\item{how_close}{set the area of trap within which the catch occurs, the default value is 0.1}
}
\description{
This function determines if lobster passes through a trap and determines if it is caught
}
