% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGrid.R
\name{updateGrid}
\alias{updateGrid}
\title{This function updates the coordinate of each lobster at each timestep,
The distanceToTrap function calculates the distance to the closest trap}
\usage{
updateGrid(
  lobsterCoordinates,
  trapCoordinates,
  trapCatch,
  lobSize,
  radius_of_influence = 15,
  dstep = 5,
  currentZoI,
  how_close = 0.1,
  q0 = 0.5,
  qmin = 0,
  saturationThreshold = 5,
  trapSaturation = T,
  lengthBased = F
)
}
\arguments{
\item{lobsterCoordinates}{is the coordinates (x,y) of each lobster at each time step}

\item{trapCoordinates}{is the coordinate of trap(s)}

\item{trapCatch}{number of lobsters caught}

\item{radius_of_influence}{is set to 15 by default}

\item{dstep}{is set to 5 by default}

\item{how_close}{is set to 0.1 by default}

\item{q0}{is set to 0.5 by default}

\item{qmin}{is set to 0 by default}

\item{trapSaturation}{is set to TRUE by default}

\item{lengthBased}{is set to FALSE by default}

\item{is}{the bait's area of influence at each timestpe}

\item{saturationThresholdis}{set to 5 by default}
}
\value{
a list of new coordinates and number of catch
}
\description{
This function updates the coordinate of each lobster at each timestep,
The distanceToTrap function calculates the distance to the closest trap
}
