% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalMove.R
\name{directionalMove}
\alias{directionalMove}
\title{This function moves the lobster a fixed distance which will be toward the trap (+some randomness) if within a zone of attraction. The closer the lobster is to the trap the smaller the random component of movement.}
\usage{
directionalMove(
  xLobster,
  yLobster,
  dStep,
  distanceToTrap,
  xtrap = 5,
  ytrap = 5,
  radius_of_influence = 15,
  ZoI
)
}
\arguments{
\item{xLobster}{location lobster in x direction}

\item{yLobster}{location of a lobster in y direction}

\item{dStep}{is how much a lobster moves in each time step}

\item{distanceToTrap}{how much influence the trap has on movement}

\item{xtrap}{location of the closest trap in the x direction}

\item{ytrap}{location of the trap in y direction}

\item{radius_of_influence}{how large is the attraction zone for the trap at t0}

\item{ZoI}{is trap's zone of influence which gets updated in each timestep}
}
\description{
This function moves the lobster a fixed distance which will be toward the trap (+some randomness) if within a zone of attraction. The closer the lobster is to the trap the smaller the random component of movement.
}
