% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceToTrapCalculator.R
\name{distanceToTrapCalculator}
\alias{distanceToTrapCalculator}
\title{This function calculates the Euclidean distance between trap(s) and each individual lobster}
\usage{
distanceToTrapCalculator(Lobster, trap = x(5, 5))
}
\arguments{
\item{Lobster}{is the matrix of lobster coordinates}

\item{trap}{is the location of the trap}
}
\description{
This function calculates the Euclidean distance between trap(s) and each individual lobster
}
