% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateLobsterMovement.R
\name{SimulateLobsterMovement}
\alias{SimulateLobsterMovement}
\title{This function moves lobsters based on the provided parameters.}
\usage{
SimulateLobsterMovement(p = p, plot = F)
}
\arguments{
\item{p}{can be modified to check the impact of each factor (i.e p$smult = 0.9 sets the shrinkage factor)}

\item{plot}{is set to false by default}
}
\description{
This function moves lobsters based on the provided parameters.
}
